Put components in separate folders
